Refactor PmStream classes to a common mother
